--- src/app/LookUpRoutineUseCaseFactory.java
@@ -1,2 +1,0 @@
-package app;public class LookUpRoutineUseCaseFactory {
-}
